refactor(#792): split meter_wiring_tests into base + signature/build_streams suite (sweep)
meter_wiring_tests.rs 618 -> 310, + meter_wiring_signature_tests.rs ~305 (both under
cap; off debt). The timer-signature + build_streams tests (self-contained, with their
own RecordingTapApi mock) move to the sibling; the base keeps ring_with. adapter-gui
lib tests (540) pass, zero warnings.
