+++ scripts/validate.sh
@@ -53,7 +53,6 @@ crates/adapter-gui/src/compact_chain_block_handlers.rs
 crates/adapter-gui/src/desktop_app.rs
 crates/adapter-gui/src/i18n_tests.rs
 crates/adapter-gui/src/lib_tests.rs
-crates/adapter-gui/src/meter_wiring_tests.rs
 crates/adapter-gui/src/project_admin_persistence_tests.rs
 crates/adapter-gui/src/project_ops_persistence_tests.rs
 crates/application/src/command.rs
